backend/requirements.txt
pyvisa>=1.13.0
pyvisa-py>=0.7.1
pandas>=2.2.2
openpyxl>=3.1.5
PyYAML>=6.0.1
